Move ImportData, RegistrationImportDatahub, KoboImportData to registration_data
