+++ backend/hct_mis_api/apps/registration_data/models.py
@@ -117,17 +117,6 @@ class RegistrationDataImport(TimeStampedUUIDModel, ConcurrencyModel, AdminUrlMix
         (EDOPOMOGA, "eDopomoga"),
     )
 
-    LOADING = "LOADING"
-    NOT_STARTED = "NOT_STARTED"
-    STARTED = "STARTED"
-    DONE = "DONE"
-    IMPORT_DONE_CHOICES = (
-        (LOADING, _("Loading")),
-        (NOT_STARTED, _("Not Started")),
-        (STARTED, _("Started")),
-        (DONE, _("Done")),
-    )
-
     name = CICharField(
         max_length=255,
         unique=True,
@@ -181,13 +170,6 @@ class RegistrationDataImport(TimeStampedUUIDModel, ConcurrencyModel, AdminUrlMix
     )
     erased = models.BooleanField(default=False, help_text="Abort RDI")
     refuse_reason = models.CharField(max_length=100, blank=True, null=True)
-    import_data = models.OneToOneField(
-        ImportData,
-        related_name="registration_data_import",
-        on_delete=models.CASCADE,
-        null=True,
-    )
-    import_done = models.CharField(max_length=15, choices=IMPORT_DONE_CHOICES, default=NOT_STARTED)
 
     def __str__(self) -> str:
         return self.name
@@ -241,3 +223,45 @@ class KoboImportedSubmission(models.Model):
         blank=True,
         on_delete=models.CASCADE,
     )
+
+
+class RegistrationDataImportDatahub(TimeStampedUUIDModel):
+    LOADING = "LOADING"
+    NOT_STARTED = "NOT_STARTED"
+    STARTED = "STARTED"
+    DONE = "DONE"
+    IMPORT_DONE_CHOICES = (
+        (LOADING, _("Loading")),
+        (NOT_STARTED, _("Not Started")),
+        (STARTED, _("Started")),
+        (DONE, _("Done")),
+    )
+
+    name = models.CharField(max_length=255, blank=True)
+    import_date = models.DateTimeField(auto_now_add=True)
+    hct_id = models.UUIDField(null=True, db_index=True)
+    import_data = models.OneToOneField(
+        ImportData,
+        related_name="registration_data_import",
+        on_delete=models.CASCADE,
+        null=True,
+    )
+    import_done = models.CharField(max_length=15, choices=IMPORT_DONE_CHOICES, default=NOT_STARTED)
+    business_area_slug = models.CharField(max_length=250, blank=True)
+
+    class Meta:
+        ordering = ("name",)
+        permissions = (["api_upload", "Can upload"],)
+
+    def __str__(self) -> str:
+        return self.name
+
+    @property
+    def business_area(self) -> str:
+        return self.business_area_slug
+
+    @property
+    def linked_rdi(self) -> "RegistrationDataImport":
+        from hct_mis_api.apps.registration_data.models import RegistrationDataImport
+
+        return RegistrationDataImport.objects.get(datahub_id=self.id)
