+++ backend/hct_mis_api/apps/registration_datahub/fixtures.py
@@ -18,13 +18,13 @@ from hct_mis_api.apps.household.models import (
     SEX_CHOICE,
     UNICEF,
 )
+from hct_mis_api.apps.registration_data.models import RegistrationDataImportDatahub
 from hct_mis_api.apps.registration_datahub.models import (
     ImportedBankAccountInfo,
     ImportedDocument,
     ImportedDocumentType,
     ImportedHousehold,
     ImportedIndividual,
-    RegistrationDataImportDatahub,
 )
 
 faker = Faker()
