+++ backend/hct_mis_api/apps/registration_datahub/schema.py
@@ -35,20 +35,22 @@ from hct_mis_api.apps.household.models import (
     ROLE_NO_ROLE,
     ROLE_PRIMARY,
 )
+from hct_mis_api.apps.registration_data.models import (
+    ImportData,
+    KoboImportData,
+    RegistrationDataImportDatahub,
+)
 from hct_mis_api.apps.registration_datahub.filters import (
     ImportedHouseholdFilter,
     ImportedIndividualFilter,
 )
 from hct_mis_api.apps.registration_datahub.models import (
-    ImportData,
     ImportedDocument,
     ImportedDocumentType,
     ImportedHousehold,
     ImportedIndividual,
     ImportedIndividualIdentity,
     ImportedIndividualRoleInHousehold,
-    KoboImportData,
-    RegistrationDataImportDatahub,
 )
 from hct_mis_api.apps.utils.schema import Arg, FlexFieldsScalar
