+++ backend/hct_mis_api/apps/registration_datahub/tasks/pull_kobo_submissions.py
@@ -10,7 +10,7 @@ from django.db import transaction
 from hct_mis_api.apps.core.kobo.api import KoboAPI
 from hct_mis_api.apps.core.kobo.common import count_population
 from hct_mis_api.apps.core.models import BusinessArea
-from hct_mis_api.apps.registration_datahub.models import KoboImportData
+from hct_mis_api.apps.registration_data.models import KoboImportData
 from hct_mis_api.apps.registration_datahub.validators import (
     KoboProjectImportDataInstanceValidator,
 )
