+++ backend/hct_mis_api/apps/registration_datahub/tasks/validatate_xlsx_import.py
@@ -6,7 +6,7 @@ from django.db import transaction
 
 import openpyxl
 
-from hct_mis_api.apps.registration_datahub.models import ImportData
+from hct_mis_api.apps.registration_data.models import ImportData
 from hct_mis_api.apps.registration_datahub.validators import UploadXLSXInstanceValidator
