+++ backend/hct_mis_api/apps/registration_datahub/tests/test_registration_xlsx_import_task.py
@@ -8,14 +8,16 @@ from django.test import TestCase
 from hct_mis_api.apps.account.fixtures import BusinessAreaFactory
 from hct_mis_api.apps.program.fixtures import ProgramFactory
 from hct_mis_api.apps.registration_data.fixtures import RegistrationDataImportFactory
-from hct_mis_api.apps.registration_data.models import RegistrationDataImport
+from hct_mis_api.apps.registration_data.models import (
+    RegistrationDataImport,
+    RegistrationDataImportDatahub,
+)
 from hct_mis_api.apps.registration_datahub.celery_tasks import (
     registration_xlsx_import_task,
 )
 from hct_mis_api.apps.registration_datahub.fixtures import (
     RegistrationDataImportDatahubFactory,
 )
-from hct_mis_api.apps.registration_datahub.models import RegistrationDataImportDatahub
 
 
 class TestRegistrationXlsxImportTask(TestCase):
