+++ backend/hct_mis_api/aurora/services/generic_registration_service.py
@@ -11,6 +11,7 @@ from hct_mis_api.apps.household.models import (
     ROLE_ALTERNATE,
     ROLE_PRIMARY,
 )
+from hct_mis_api.apps.registration_data.models import RegistrationDataImportDatahub
 from hct_mis_api.apps.registration_datahub.models import (
     ImportedBankAccountInfo,
     ImportedDocument,
@@ -18,7 +19,6 @@ from hct_mis_api.apps.registration_datahub.models import (
     ImportedHousehold,
     ImportedIndividual,
     ImportedIndividualRoleInHousehold,
-    RegistrationDataImportDatahub,
 )
 from hct_mis_api.aurora.services.base_flex_registration_service import (
     BaseRegistrationService,
