+++ backend/hct_mis_api/aurora/services/sri_lanka_flex_registration_service.py
@@ -17,6 +17,7 @@ from hct_mis_api.apps.household.models import (
     ROLE_PRIMARY,
     YES,
 )
+from hct_mis_api.apps.registration_data.models import RegistrationDataImportDatahub
 from hct_mis_api.apps.registration_datahub.models import (
     ImportedBankAccountInfo,
     ImportedDocument,
@@ -24,7 +25,6 @@ from hct_mis_api.apps.registration_datahub.models import (
     ImportedHousehold,
     ImportedIndividual,
     ImportedIndividualRoleInHousehold,
-    RegistrationDataImportDatahub,
 )
 from hct_mis_api.apps.utils.age_at_registration import calculate_age_at_registration
 from hct_mis_api.aurora.services.base_flex_registration_service import (
